:truck: Rename ui/ as controlnet_ui/
