+++ scripts/controlnet_ui/controlnet_ui_group.py
